refactor: 앱단 서비스 분리
